Improves board styling and refactors board sizing state setup
