+++ src/css/Header.sass
@@ -1,5 +1,4 @@
 header
-    background-color: #FFFFFF
     display: flex
     align-items: center
     justify-content: space-between
