+++ src/css/LifeBoard.sass
@@ -1,8 +1,9 @@
-.board-wrapper
-    margin: 0 33px
-    display: flex
-    justify-content: center
-    height: calc(100vh - 100px)
-    
 .life-board
-    border-radius: 3px
+    margin: 0 2rem
+    height: calc(100vh - 100px)
+    display: flex
+    flex-wrap: wrap
+    justify-content: center
+    align-items: center
+    align-content: flex-start
+
