+++ src/index.sass
@@ -12,6 +12,7 @@ html
     height: 100%
     color: $text
     overflow: hidden
+    background-color: #FFFAFA
 
 body
     margin: 0
