+++ src/utils/types.ts
@@ -2,3 +2,8 @@ export interface Cell {
     cellStatus: number;
     id: number;
 }
+
+export interface BoardDimensions {
+    width: number;
+    height: number;
+}
